Struct fedimint_client_wasm::WasmClient

pub struct WasmClient {
    pub(crate) client: ClientHandleArc,
}

Fields
client: ClientHandleArc
Implementations

impl WasmClient

pub async fn open(client_name: String) -> Result<Option<WasmClient>, JsError>

Open fedimint client with already joined federation.

After you have joined a federation, you can reopen the fedimint client with same client_name. Opening client with same name at same time is not supported. You can close the current client by calling client.free(). NOTE: The client will remain active until all the running rpc calls have finished.

pub async fn join_federation( client_name: String, invite_code: String, ) -> Result<WasmClient, JsError>

Open a fedimint client by join a federation.

pub(crate) async fn join_federation_inner( invite_code: String, ) -> Result<WasmClient>

pub async fn rpc( &self, module: &str, method: &str, payload: String, cb: &Function, )

Call a fedimint client rpc the responses are returned using cb callback. Each rpc call can return multiple responses by calling cb multiple times. You should ignore the promise by this function because it has no significance.

pub(crate) fn rpc_inner<'a>( &'a self, module: &'a str, method: &'a str, payload: String, ) -> impl Stream<Item = Result<Value>> + 'a